requirements.txt
pandas
scikit-learn
numpy
matplotlib
skop